First submission of file from Bloomberg
This is the original Bloomberg file copied from the Bberg Terminal. This file implements Black-Litterman as designed by Bberg taking in a portfolio computing the implied returns and then incorporating our views outputs a result. Goal is to substitute the commands that use data from Bberg's database to those that take data from Excel.