Changed the display of additional and unnecessary sliders
We commented out a line that forcefully displayed sliders. This line was left in from previous attempts to have sliders dislayed that were not successful due to the wrong implementation of bqplot. As a result, now that the rest of the interactive bqplot requiring items are displayed the ended up displaying additional and unwanted sliders.