The code is running very well and as expected in this current configuration.
When run the code will output good and logical results. It assigns the majority of the portfolio to sure and stable assets such as bonds. If the return of those is then taken to negative territory with the slider controlling the views, the new Black Litterman weights will respond logically and assign practically zero portion of the portfolio to those assets.