Applied new input to the code.
In this version we supply as input to our code a new data file, originally called "TEA-...". We also re-arrange our input files away from the Main folder into the archive folder to avoid cluttering of our main workspace. We created a new data folder called "Cleaned" which will contani our data in cleaned form such that we can then simply copy the file that we want to the "Main" directory, rename it to "prices" (which is the input for which our code looks) and then simply run it.