deleting irrelevant file
We delete a file that was wrongly commted previously. Basically, we previously commited a file that was the file that word creates when you open a word document and keeps open while your word file is open. It then deletes it once you save and close your file. I had accidentally commited that here.