Understood the workings of the float slider
To type in a value you are required to type the value you like by dividing it in your mind by 100. Meaning that if you want to input a "view" of 5% you need to type in 0.05. I would like to search where exactly the values are read in and then divide those by 100 so that the user will be able to input 5 and the code to understand 0.05 in reality