Changing the Risk-free rate
We changed the risk free rate to zero to reflect current economic conditions. The initial risk-free rate was set by the US team developing this code where the economic enviroment reflected such an estimation. On the contrary in the EU with the pandemic environment  we need to lower it significantly down to zero.