Reporting Commit: Problem located
Problem was found to be the changing of the risk-free rate. By reducing the risk-free rate to zero the portfolio is then unable to reach the return target of 4% that we have set. As a result, the program breaks down leading to a wrong and unresponsive graph in the end.