Solution for previous problem.
As explained in the previous commit we got an unresponsive graph due to a problem in the code. Namely, by reducing the risk free rate down to zero to reflect the current economic environment our portfolio was then unable to reach the 4% target for the return. By reducing the return target down to 1% the portfolio is now able to reach that target and hence the code does not collapse resulting in a responsive graph, as expected.