Suppressing commit 2
We suppressed a big print that is done by the solve_intial_opt_weight function. That function is simply called but in reality does not need to return anything since its job is to just be called. As a result we comment out its return line to suppress the unnecessary print outs.