Simplifying the product notebook
We dumb down the notebook that will be displayed to clients. Clients are unable to specifically run the code so we hide code and place a button that automatically runs all cells when clicked on. This way the client will not have to specifically run all cell on their own.

We still need to link the initial weights sliders to the code.